+++ src/lib.rs
@@ -6,3 +6,4 @@
 
 pub mod command;
 pub mod display;
+pub mod size;