+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Migrate to embedded-graphics-core 0.3 (embedded-graphics-0.7 API).
  - Migrate examples to embedded-graphics 0.7 API.
  - Add bounding box check
+ - Add support for different-sized displays.
 
 ## v0.1.0 - 2020-11-25